# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hemansh Manager API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskCategory = (typeof TaskCategory)[keyof typeof TaskCategory];

export const TaskCategory = {
  Study: "Study",
  Gym: "Gym",
  Personal: "Personal",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  High: "High",
  Medium: "Medium",
  Low: "Low",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  category: TaskCategory;
  priority: TaskPriority;
  completed: boolean;
  deadline?: string | null;
  date?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyCategory =
  (typeof CreateTaskBodyCategory)[keyof typeof CreateTaskBodyCategory];

export const CreateTaskBodyCategory = {
  Study: "Study",
  Gym: "Gym",
  Personal: "Personal",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  High: "High",
  Medium: "Medium",
  Low: "Low",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string | null;
  category: CreateTaskBodyCategory;
  priority: CreateTaskBodyPriority;
  deadline?: string | null;
  date?: string | null;
}

export type UpdateTaskBodyCategory =
  (typeof UpdateTaskBodyCategory)[keyof typeof UpdateTaskBodyCategory];

export const UpdateTaskBodyCategory = {
  Study: "Study",
  Gym: "Gym",
  Personal: "Personal",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  High: "High",
  Medium: "Medium",
  Low: "Low",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  category?: UpdateTaskBodyCategory;
  priority?: UpdateTaskBodyPriority;
  completed?: boolean;
  deadline?: string | null;
  date?: string | null;
}

export interface StudySession {
  id: number;
  subject: string;
  hours: number;
  date: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateStudySessionBody {
  subject: string;
  hours: number;
  date: string;
  notes?: string | null;
}

export interface UpdateStudySessionBody {
  subject?: string;
  hours?: number;
  date?: string;
  notes?: string | null;
}

export type StudyAnalyticsSubjectBreakdownItem = {
  subject: string;
  hours: number;
};

export type StudyAnalyticsDailyHoursItem = {
  date: string;
  hours: number;
};

export interface StudyAnalytics {
  currentStreak: number;
  totalHoursThisWeek: number;
  totalHoursToday: number;
  subjectBreakdown: StudyAnalyticsSubjectBreakdownItem[];
  dailyHours: StudyAnalyticsDailyHoursItem[];
}

export interface Exercise {
  name: string;
  sets?: number | null;
  reps?: number | null;
  weight?: number | null;
}

export interface GymSession {
  id: number;
  workoutType: string;
  exercises: Exercise[];
  date: string;
  durationMinutes?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateGymSessionBody {
  workoutType: string;
  exercises: Exercise[];
  date: string;
  durationMinutes?: number | null;
  notes?: string | null;
}

export interface UpdateGymSessionBody {
  workoutType?: string;
  exercises?: Exercise[];
  date?: string;
  durationMinutes?: number | null;
  notes?: string | null;
}

export type GymAnalyticsDailyActivityItem = {
  date: string;
  didWorkout: boolean;
};

export interface GymAnalytics {
  currentStreak: number;
  sessionsThisWeek: number;
  sessionsThisMonth: number;
  dailyActivity: GymAnalyticsDailyActivityItem[];
}

export interface Note {
  id: number;
  title: string;
  content: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  title: string;
  content: string;
  tags?: string[];
}

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  tags?: string[];
}

export interface DashboardSummary {
  date: string;
  tasksTotal: number;
  tasksCompleted: number;
  studyHoursToday: number;
  studyGoalHours: number;
  didGymToday: boolean;
  gymStreak: number;
  studyStreak: number;
  upcomingTasks: Task[];
}

export interface MotivationalQuote {
  quote: string;
  author: string;
}

export interface Conversation {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type AiMessageRole = (typeof AiMessageRole)[keyof typeof AiMessageRole];

export const AiMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AiMessage {
  id: string;
  role: AiMessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: string;
  title: string;
  messages: AiMessage[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateConversationBody {
  title: string;
}

export interface SendAiMessageBody {
  content: string;
}

export interface AiSuggestions {
  suggestions: string[];
  generatedAt: string;
}

export type ListTasksParams = {
  category?: ListTasksCategory;
  priority?: ListTasksPriority;
  completed?: boolean;
  date?: string;
};

export type ListTasksCategory =
  (typeof ListTasksCategory)[keyof typeof ListTasksCategory];

export const ListTasksCategory = {
  Study: "Study",
  Gym: "Gym",
  Personal: "Personal",
} as const;

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  High: "High",
  Medium: "Medium",
  Low: "Low",
} as const;

export type ListStudySessionsParams = {
  date?: string;
  week?: string;
};

export type ListGymSessionsParams = {
  date?: string;
};

export type ListNotesParams = {
  tag?: string;
  search?: string;
};