# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hemansh Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gym/sessions:
    get:
      operationId: listGymSessions
      tags: [gym]
      summary: List gym sessions
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of gym sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GymSession"
    post:
      operationId: createGymSession
      tags: [gym]
      summary: Log a gym session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGymSessionBody"
      responses:
        "201":
          description: Created gym session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymSession"
  /gym/sessions/{id}:
    patch:
      operationId: updateGymSession
      tags: [gym]
      summary: Update a gym session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGymSessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymSession"
    delete:
      operationId: deleteGymSession
      tags: [gym]
      summary: Delete a gym session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DASHBOARD ─────────────────────────────────────────────────────────
...
  /ai/conversations:
    get:
      operationId: listConversations
      tags: [ai]
      summary: List AI conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [ai]
      summary: Create a new AI conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /ai/conversations/{id}:
    get:
      operationId: getConversation
      tags: [ai]
      summary: Get a conversation with its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
    delete:
      operationId: deleteConversation
      tags: [ai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /ai/conversations/{id}/messages:
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send a message in a conversation (streaming SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAiMessageBody"
      responses:
        "200":
          description: Streaming SSE response
...
components:
  schemas:
...
    GymSession:
      type: object
      properties:
        id:
          type: integer
        workoutType:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
        date:
          type: string
          format: date
        durationMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, workoutType, exercises, date, createdAt]
    Exercise:
      type: object
      properties:
        name:
          type: string
        sets:
          type: integer
          nullable: true
        reps:
          type: integer
          nullable: true
        weight:
          type: number
          nullable: true
      required: [name]
    CreateGymSessionBody:
      type: object
      properties:
        workoutType:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
        date:
          type: string
          format: date
        durationMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [workoutType, exercises, date]
    UpdateGymSessionBody:
      type: object
      properties:
        workoutType:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
        date:
          type: string
          format: date
        durationMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
...
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, content, tags, createdAt, updatedAt]
    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, content]
    UpdateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── Dashboard ──
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, createdAt, updatedAt]
    ConversationWithMessages:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiMessage"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, messages, createdAt, updatedAt]
    AiMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, role, content, createdAt]
    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
      required: [title]
    SendAiMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
...