# lib/api-zod/src/generated/types/updateGymSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hemansh Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";

export interface UpdateGymSessionBody {
  workoutType?: string;
  exercises?: Exercise[];
  date?: Date;
  durationMinutes?: number | null;
  notes?: string | null;
}